refactor(template): 메일 발송 테스트 코드 수정
