+++ mail-client/src/main/resources/templates/mail-template.html
@@ -134,13 +134,7 @@
               <tr>
                 <td
                   style="width: 100%; padding: 0 0 10px 0; text-align: center; color: #808080; font-size: 0.8rem">
-                  위 버튼을 클릭하시거나 다음 링크를 열으세요:<br/>
-                  <span th: text="${redirectUrl}"></span>
-                </td>
-              </tr>
-              <tr>
-                <td
-                  style="width: 100%; padding: 0 0 20px 0; text-align: center; color: #808080; font-size: 0.8rem">
+                  위 버튼을 클릭하시거나 다음 <a th:href="${redirectUrl}">링크</a>를 열으세요<br/>
                   위 링크는 24시간동안 유효합니다.
                 </td>
               </tr>
