+++ mail-client/src/test/kotlin/com/devfloor/hongit/client/mail/domain/impl/ProdMailSenderTest.kt
@@ -1,12 +1,10 @@
-package com.devfloor.hongit.api.auth.application
+package com.devfloor.hongit.client.mail.domain.impl
 
-import com.devfloor.hongit.api.auth.application.request.AuthMailRequest
-import com.devfloor.hongit.api.support.MockitoHelper.any
 import com.devfloor.hongit.client.mail.common.config.MailSenderConfig
 import com.devfloor.hongit.client.mail.domain.spec.MailSender
+import com.devfloor.hongit.client.mail.support.MockitoHelper.any
 import com.devfloor.hongit.core.authtoken.AuthToken
 import com.devfloor.hongit.core.authtoken.AuthTokenRepository
-import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
@@ -18,37 +16,24 @@ import org.springframework.test.context.ActiveProfiles
 import org.springframework.test.context.junit.jupiter.SpringExtension
 import org.thymeleaf.spring5.SpringTemplateEngine
 
-/**
- * 실제 메시지 전송 여부를 테스트하기 위한 테스트 코드 (profile = prod)
- * 테스트 시 @Disable 어노테이션 주석처리 후 실행
- */
 @Disabled
 @ActiveProfiles(value = ["prod"])
 @ExtendWith(SpringExtension::class)
-@SpringBootTest(
-    classes = [AuthService::class, MailSenderConfig::class, SpringTemplateEngine::class],
-    properties = ["mail.sender", "hongit"]
-)
-internal class ProdAuthServiceTest {
-    private lateinit var service: AuthService
-
+@SpringBootTest(classes = [MailSenderConfig::class, SpringTemplateEngine::class])
+internal class ProdMailSenderTest {
     @Autowired
     private lateinit var mailSender: MailSender
 
     @MockBean
     private lateinit var authTokenRepository: AuthTokenRepository
 
-    @BeforeEach
-    internal fun setUp() {
-        service = AuthService(authTokenRepository, mailSender)
-    }
-
     @Test
-    internal fun `sendAuthenticationMail - 인증 토큰에 따른 메일 발송`() {
+    internal fun `send - 인증메일 발송`() {
         // given
-        given(authTokenRepository.save(any())).willReturn(AuthToken())
+        val authToken = AuthToken()
+        given(authTokenRepository.save(any())).willReturn(authToken)
 
         // when
-        service.sendAuthenticationMail(AuthMailRequest("ljy3510@gmail.com"))
+        mailSender.send(mailSender.createMimeMessage("ljy3510@gmail.com", authToken.id.toString()))
     }
 }
