+++ mail-client/src/test/kotlin/com/devfloor/hongit/client/mail/support/MockitoHelper.kt
@@ -1,0 +1,7 @@
+package com.devfloor.hongit.client.mail.support
+
+import org.mockito.Mockito
+
+object MockitoHelper {
+    inline fun <reified T> any(): T = Mockito.any(T::class.java)
+}
